requirements.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python
numpy
websockets
streamlit
pandas
pillow